+++ planner/core/casetest/BUILD.bazel
@@ -5,9 +5,7 @@ go_test(
     timeout = "moderate",
     srcs = [
         "binary_plan_test.go",
-        "enforce_mpp_test.go",
         "expression_rewriter_test.go",
-        "flat_plan_test.go",
         "integration_partition_test.go",
         "integration_test.go",
         "main_test.go",
@@ -32,14 +30,11 @@ go_test(
         "//domain",
         "//expression",
         "//expression/aggregation",
-        "//infoschema",
-        "//kv",
         "//parser",
         "//parser/ast",
         "//parser/model",
         "//parser/mysql",
         "//parser/terror",
-        "//planner",
         "//planner/core",
         "//planner/core/internal",
         "//planner/property",
@@ -48,13 +43,11 @@ go_test(
         "//sessionctx/variable",
         "//statistics/handle",
         "//testkit",
-        "//testkit/external",
         "//testkit/testdata",
         "//testkit/testmain",
         "//testkit/testsetup",
         "//types",
         "//util",
-        "//util/collate",
         "//util/hint",
         "//util/logutil",
         "//util/mock",
