+++ planner/core/casetest/enforcempp/BUILD.bazel
@@ -1,0 +1,27 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "enforcempp_test",
+    timeout = "short",
+    srcs = [
+        "enforce_mpp_test.go",
+        "main_test.go",
+    ],
+    data = glob(["testdata/**"]),
+    flaky = True,
+    shard_count = 12,
+    deps = [
+        "//domain",
+        "//parser/model",
+        "//planner/core/internal",
+        "//sessionctx/stmtctx",
+        "//testkit",
+        "//testkit/external",
+        "//testkit/testdata",
+        "//testkit/testmain",
+        "//testkit/testsetup",
+        "//util/collate",
+        "@com_github_stretchr_testify//require",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
